+++ CHANGELOG.md
@@ -57,6 +57,9 @@ All notable changes to this project will be documented in this file.
 - Add client export for GetPlayersPed function
 - Add client export for GetPlayersId function
 - Add client export for GetPlayersPedOrderById function
+- Add client export for Clone function
+- Add client export for Copy function
+- Add server export for Copy function
 - Add server export for SetDebug function
 
 ### Changed